libSM: update to 1.2.6.
Support for using getaddrinfo() was originally added to support IPv6,
and only used if IPv6 support was enabled.  Two decades later, support
for getaddrinfo() is ubiquitous and OS'es have marked gethostbyname()
as deprecated, so use the modern interface whenever we can now.
(Note that code included from xtrans will only use the modern interfaces
when IPv6 is disabled if building against xtrans 1.6.0 or later.)

Of course, in libSM, this only matters if one of the UUID generation
interfaces isn't being used instead, and the configure script has also
been modified in this release to stop checking for network interfaces
if we're not going to use them because we found a UUID generation
mechanism to use instead for creating ids.

Alan Coopersmith (6):
      Skip network checks/headers/libraries if using UUID-based ids
      sm_genid.c: simplify final fallback case
      Simplify ifdefs for IPv6 support
      if getaddrinfo() is available, use it, even if IPv6 support is disabled
      SmsGenerateClientID: simplify code to convert IPv4 address to hex string
      libSM 1.2.6