py-hypothesis: updated to 6.129.0
6.129.0

This release adds a "hypothesis-urandom" backend, which draws randomness from /dev/urandom instead of Python’s PRNG. This is useful for users of Antithesis who also have Hypothesis tests, allowing Antithesis mutation of /dev/urandom to drive Hypothesis generation. We expect it to be strictly slower than the default backend for everyone else.

It can be enabled with @settings(backend="hypothesis-urandom").