sysutils/qemu-guest-agent: New package for just the qemu-ga program.
This is an agent designed to be used inside a guest -- so it's not
really useful for cases where you want the rest of the emulators/qemu
package, and at a few megabytes it's much smaller than emulators/qemu
which weighs in at a few gigabytes.

If we had subpackages -- multiple binary packages built from a single
source -- then it would be reasonable to use that for the qemu guest
agent rather than maintaining a separate makefile.  But we don't.

Derived from wip/qemu-guest-agent.