ghostscript-agpl: updated to 10.05.0
10.05.0

This release addresses CVEs: CVE-2025-27835, CVE-2025-27832, CVE-2025-27831, CVE-2025-27836, CVE-2025-27830, CVE-2025-27833, CVE-2025-27837, CVE-2025-27834

In addition one other security fix for which a CVE is pending which will be added to the online version of this document when assigned: News

The 10.05.0 release deprecates the non-standard operator "selectdevice", all code should now be using the standard "setpagedevice" operator. "selectdevice" will be removed in the 10.06.0 release.

We now support production of PDF/X-1a and PDF/X-4a in addition to the existing support for PDF/X-3

IMPORTANT: In the 10.04.0 release we added protection for device selection from PostScript input. This will mean that, by default, only the device specified on the command line will be permitted. Similar to the file permissions, there will be a "--permit-devices=" allowing a comma separation list of allowed devices. This will also take a single wildcard "*" allowing any device.

Any application which relies on allowing PostScript to change devices during a job will have to be aware, and take action to deal with this change.

The exception is "nulldevice", switching to that requires no special action.

Our efforts in code hygiene and maintainability continue.

The usual round of bug fixes, compatibility changes, and incremental improvements.

(9.53.0) We have added the capability to build with the Tesseract OCR engine. In such a build, new devices are available (pdfocr8/pdfocr24/pdfocr32) which render the output file to an image, OCR that image, and output the image "wrapped" up as a PDF file, with the OCR generated text information included as "invisible" text (in PDF terms, text rendering mode 3).

Mainly due to time constraints, we only support including Tesseract from source included in our release packages, and not linking to Tesseract/Leptonica shared libraries. Whether we add this capability will be largely dependent on community demand for the feature.