sqlite3: Remove jimtcl dependency.
The sqlite3 autoconf distribution includes a bundled jimsh0 that is used to
bootstrap in the case of the host not providing a TCL implementation.  Simply
use that rather than depending on jimtcl, which is non-portable.

Fixes build on at least Solaris/illumos.  Tested on SmartOS, macOS, and NetBSD,
with macOS being of particular interest as it ships a native implementation of
tclsh so does not use the bootstrap version.

While here perform pkglint cleanup.  Bump PKGREVISION.