perl: update to 5.40.2.
[CVE-2024-56406] Heap buffer overflow vulnerability with tr//

A heap buffer overflow vulnerability was discovered in Perl.

When there are non-ASCII bytes in the left-hand-side of the C<tr> operator,
C<S_do_trans_invmap()> can overflow the destination pointer C<d>.

  $ perl -e '$_ = "\x{FF}" x 1000000; tr/\xFF/\x{100}/;'
  Segmentation fault (core dumped)

It is believed that this vulnerability can enable Denial of Service or
Arbitrary Code Execution attacks on platforms that lack sufficient defenses.

Discovered by: Nathan Mills.

<Module::CoreList> has been upgraded from version 5.20250118_40 to 5.20250413_40.