lang/tcl: Adhere to new separator in TCL_PACKAGE_PATH from tcl 8.6.15.
As the change log for 8.6.15 says:
* TCL_PACKAGE_PATH in tclConfig.sh change from TCL list to ":"
        (UNIX) or ";" (Windows) separated items (nijtmans)
        *** POTENTIAL INCOMPATIBILITY ***

And, indeed, yes, we were hit; tcl-scotty stopped working with the
update from 8.6.14 to 8.6.16, indicating that it could not find
the "tnm 2.1.11" extension.  TCL_PACKAGE_PATH is set in the tcl
package's unix/configure by one of our patches, and a space needed
to be a colon.

Bump PKGREVISION.